give option for static OR dynamic
adds an option to set enableForceStaticIP which when true will always
present the static IP fields - if no static IP is sent these will be
blank. If the user doesn’t fill these in then dhcp will be used. Also
added “getSTAIsStaticIP()” - call this after the save is completed and
query whether the user has set a static IP. This way the sketch can
then save the static IP to EEPROM if required.