Changed true and false in json formatted response to /state http request to be of type boolean from previous type string.
Moved scan for WiFi networks to a seperate function. Uses the third option from https://www.eskimo.com/~scs/cclass/int/sx5.html to return an array which is messy as it requires the use of pointers, malloc and free. The function returns the number of WiFi networks it found and changes a pointer to point to an array of indexes that is used to access the WiFi networks that were found array in signal strength order.

Changed the ConfigOnSwitch example to use waitForConnectedResult method rather than poll in a loop for the WiFi status when attempting to connect to WiFi.